New menu system actually working this time
