+++ Assets/Scripts/Lobby/LobbyMenuManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bd980b0b3c4f0664dbb55643ad9aeb92
+guid: fcc5f397c41308749b28917ec6e4f24e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
